refactor: reorg files application, console, color and logger
