+++ packages/core/src/application/application-factory.ts
@@ -21,6 +21,7 @@ function isWebServerConstructor<T extends IWebServer>(
  * AppFactory Class
  *
  * Responsible for creating an instance of the IWebServer implementation using a custom application type.
+ * @public API
  */
 export class AppFactory {
   public static container: Container;
