+++ packages/core/src/application/application.types.ts
@@ -2,6 +2,7 @@
  * Enum representing possible server environments.
  * @options Development - Development environment.
  * @options Production - Production environment.
+ * @public API
  */
 export enum ServerEnvironment {
   Development = "development",
