+++ packages/core/src/application/index.ts
@@ -1,2 +1,3 @@
 export { AppContainer } from "./application-container";
 export { AppFactory } from "./application-factory";
+export { ServerEnvironment } from "./application.types";
