+++ packages/core/src/common/index.ts
@@ -1,2 +1,1 @@
 export { Pattern, ExpressoConfig } from "./project-config";
-export { ServerEnvironment } from "./server-env.types";
